backend/requirements.txt
# ── Core Framework ────────────────────────────────────────────────
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# ── LLM Providers ────────────────────────────────────────────────
google-genai>=1.0.0
openai>=1.3.0

# ── Embeddings & ML ──────────────────────────────────────────────
sentence-transformers>=3.0.0
scikit-learn>=1.3.2
numpy>=1.24.0

# ── Vector Database ──────────────────────────────────────────────
pinecone>=5.0.0

# ── Graph Database ───────────────────────────────────────────────
neo4j>=5.14.0

# ── Document Processing ──────────────────────────────────────────
pypdf>=3.17.0
python-docx>=1.1.0

# ── HTTP Client (for Ollama health checks) ────────────────────────
requests>=2.31.0

# ── Testing ──────────────────────────────────────────────────────
pytest>=7.4.0
httpx>=0.25.0
